+++ wordle/conf.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+ROOT_DIR = Path(__file__).parent
+STATIC_DIR = ROOT_DIR.joinpath("static")
+TEMPLATES_DIR = ROOT_DIR.joinpath("templates")
